+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-10 09:59:34
-// Commit: refactor(core): extract reusable atomic widgets from core
+// Updated at: 2026-04-10 20:49:17
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '264';
-  final String lastMessage = 'refactor(core): extract reusable atomic widgets from core';
-  final DateTime timestamp = DateTime.parse('2026-04-10T09:59:34+05:00');
+  final String commitRef = '301';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-04-10T20:49:17+05:00');
 }